create environmentVars template
